+++ examples/companion_radio/UITask.h
@@ -46,7 +46,7 @@ class UITask {
 
   void renderCurrScreen();
   void userLedHandler();
-  void renderBatteryIndicator(uint16_t batteryPercentage);
+  void renderBatteryIndicator(int batteryPercentage);
   
   // Button action handlers
   void handleButtonAnyPress();
